fixing reducer
