+++ client/src/components/Joke.js
@@ -5,7 +5,7 @@ const Joke = ({ item }) => {
     return(
      <div key={item.id} className="tenJokes">
         <p>{item.setup}</p>
-        <p>{item.punchline}</p>
+        <p>Answers: {item.punchline}</p>
     </div>)
 
 }
